more precise features

+++ StitchMan/Keypoint.h
@@ -11,11 +11,11 @@
 @interface Keypoint : NSObject
 {
     @public
-    int x;
-    int y;
+    double x;
+    double y;
     
-    int x_oct;
-    int y_oct;
+    double x_oct;
+    double y_oct;
     
     int octave_num;
     int interval_num;
@@ -27,7 +27,7 @@
     
 }
 
-- (id)initWithX:(int)x Y:(int)y X_OCT:(int)x_oct Y_OCT:(int)y_oct
+- (id)initWithX:(double)x Y:(double)y X_OCT:(double)x_oct Y_OCT:(double)y_oct
          Octave:(int)octave_num Interval:(int)interval_num Subinterval:(double)sub_interval_num;
 
 - (id)initWithKeypoint:(Keypoint *)keypoint;
